% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelgesis-package.r
\docType{package}
\name{excelgesis-package}
\alias{excelgesis}
\alias{excelgesis-package}
\title{excelgesis: Access and Edit XML Inside xlsx}
\description{
Access and Edit XML inside xlsx. Access and Edit XML inside xlsx.
Access and Edit XML inside xlsx.
}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@stat.ubc.ca}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
