% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_browse.R
\name{xg_browse}
\alias{xg_browse}
\title{Browse around an unpacked xlsx}
\usage{
xg_browse(path)
}
\arguments{
\item{path}{Path to a directory containing an unpacked, linkified \code{.xlsx} or
to an \code{index.html} within.}
}
\value{
The input path, invisibly.
}
\description{
Browse around an unpacked xlsx
}
\examples{
file.copy(xg_example("datasets"), "datasets.xlsx")
"datasets.xlsx" \%>\%
  xg_unzip() \%>\%
  xg_linkify() \%>\%
  xg_browse()

## clean up
unlink("datasets", recursive = TRUE)
file.remove("datasets.xlsx")
}
