% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_example.R
\name{xg_example}
\alias{xg_example}
\title{Get path to excelgesis example}
\usage{
xg_example(x = character())
}
\arguments{
\item{x}{Character. Each element should be a filename with extension
\code{.xlsx} extension or a regex that matches desired filename(s).}
}
\value{
Character vector of full file paths.
}
\description{
excelgesis includes several sheets drawn from examples, tests, and issues on
readxl, rexcel, and googlesheets. This function makes it easier to get these
filepaths.
}
\examples{
xg_example()
xg_example(c("iris-excel-xlsx.xlsx", "datasets.xlsx"))
xg_example("missing")
xg_example(c("iris", "missing"))
xg_example(c("missing", "datasets.xlsx"))
}
