% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_unzip.R
\name{xg_unzip}
\alias{xg_unzip}
\title{Unzip an xlsx}
\usage{
xg_unzip(path, exdir = character())
}
\arguments{
\item{path}{Character, length one. Path to xlsx file.}

\item{exdir}{Passed to \code{tools::unzip}. Directory in which to unpack the
xlsx.}
}
\value{
The path to the top-level directory of the unpacked xlsx, invisibly.
}
\description{
Extracts the files from an xlsx zip archive. Note it will overwrite existing
files, so plan accordingly.
}
\examples{
(path <- xg_example("mini"))
(x <- xg_unzip(path))
}
