% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_example.R
\name{xg_example}
\alias{xg_example}
\title{Get path to excelgesis example}
\usage{
xg_example(x = NULL)
}
\arguments{
\item{x}{Character. A regex that matches desired filename(s).}
}
\value{
Character vector of full file paths.
}
\description{
excelgesis includes many sheets drawn from examples, tests, and issues on
readxl, rexcel, and googlesheets. This function makes it easier to get these
filepaths. You probably want to make a copy in, e.g., working directory
before unpacking an example. Or specify the \code{exdir} in \code{xg_unzip()}.
}
\examples{
xg_example()
xg_example("iris-excel-xlsx.xlsx")
xg_example("datasets")
xg_example("missing")
}
