% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_unzip.R
\name{xg_unzip}
\alias{xg_unzip}
\title{Unzip an xlsx}
\usage{
xg_unzip(path, exdir = NULL)
}
\arguments{
\item{path}{Character, length one. Path to xlsx file.}

\item{exdir}{Passed to \code{\link[utils:unzip]{utils::unzip()}}. Directory in which to unpack the
xlsx. By default, if unpacking \code{foo.xlxs}, this will be \code{foo/} in
current working directory.}
}
\value{
The path to the top-level directory of the unpacked xlsx, invisibly.
}
\description{
Extracts the files from an xlsx zip archive. Note it will overwrite existing
files, so plan accordingly.
}
\examples{
(path <- xg_example("datasets"))
xg_unzip(path)
}
